Address minor comments, move Topic into o.a.k.common.internals
