+++ clients/src/test/java/org/apache/kafka/common/internals/TopicTest.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.apache.kafka.common;
+package org.apache.kafka.common.internals;
 
 import org.junit.Test;
