+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/InternalTopicConfig.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.streams.processor.internals;
 
-import org.apache.kafka.common.Topic;
+import org.apache.kafka.common.internals.Topic;
 
 import java.util.Map;
 import java.util.Objects;
